requirements.txt
pymongo
psutil
Faker
PyYAML